Cloud Computing/Machine Learning Deployment/model-prediction/farmacode-function/requirements.txt
numpy>=1.18.1
tensorflow
google-cloud-storage>=1.38.0
opencv-python
firebase-admin
Pillow